requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
PyMuPDF
transformers
torch
sentencepiece 
streamlit 